+++ src/Validator.js
@@ -3,6 +3,13 @@ import PropTypes from "prop-types";
 import Field from "./field";
 import validate from "./validate-fns";
 
+import {
+  LengthValidation,
+  MatchValidation,
+  EqualsValidation,
+  RequiredValidation
+} from "./validations";
+
 export class Validator extends React.Component {
   fields = {};
   values = {};
@@ -13,8 +20,14 @@ export class Validator extends React.Component {
   };
   constructor(props) {
     super(props);
-    let { validators } = props;
-    this.validators = validators;
+    let { validators = [] } = props;
+    this.validators = [
+      LengthValidation,
+      MatchValidation,
+      EqualsValidation,
+      RequiredValidation,
+      ...validators
+    ];
     this.register = this.register.bind(this);
     this.validate = this.validate.bind(this);
     this.onChange = this.onChange.bind(this);
